PET-30 Edit pet
Edit pet now works, sort of.
Still some workarounds and fixes to do, but on a prototype level it's passable.
Bug hunting took a while, but the problem was quite simple. JSON was formatted correctly but the PetId on the URL was allways 0 because it was never defined to be the same as the pet we were looking at.